Move withoutStyleProps to helpers
