+++ packages/palette/src/helpers/withoutStyleProps.tsx
